Add Augmenter Class
