+++ content/augmenter.js
@@ -1,0 +1,78 @@
+// Copyright 2015, 2016 Parity Technologies (UK) Ltd.
+// This file is part of Parity.
+
+// Parity is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+
+// Parity is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+
+// You should have received a copy of the GNU General Public License
+// along with Parity.  If not, see <http://www.gnu.org/licenses/>.
+
+export default class Augmenter {
+
+  static augmentNode (key, node, resolved = {}) {
+    if (!node || node.getAttribute('data-parity-touched') === 'true') {
+      return;
+    }
+
+    node.setAttribute('data-parity-touched', true);
+
+    if (!resolved[key]) {
+      return;
+    }
+
+    const { address } = resolved[key];
+
+    node.outerHTML += `<span data-parity-touched="true"> (${address})</span>`;
+  }
+
+  static run (matches, resolved = {}) {
+    // Use the attributes matcher first
+    const attributesMatches = matches.filter((match) => match.from === 'attributes');
+    const textMatches = matches.filter((match) => match.from === 'text');
+
+    attributesMatches
+      .forEach((match) => {
+        const { email, node } = match;
+        Augmenter.augmentNode(email, node, resolved);
+      });
+
+    textMatches
+      .forEach((match) => {
+        const { email, node } = match;
+
+        // Safe Node is if the node which inner text is only the email address
+        let safeNode = node.innerText.trim() === email
+          ? node
+          : null;
+
+        // If it has more text, try to separate in SPANs
+        if (!safeNode) {
+          const emailIndex = node.innerText.indexOf(email);
+
+          if (emailIndex === -1) {
+            return;
+          }
+
+          const beforeText = node.innerText.slice(0, emailIndex);
+          const afterText = node.innerText.slice(emailIndex + email.length);
+
+          node.innerHTML = `${beforeText}<span>${email}</span>${afterText}`;
+          safeNode = node.querySelector('span');
+        }
+
+        if (!safeNode) {
+          return;
+        }
+
+        Augmenter.augmentNode(email, safeNode, resolved);
+      });
+  }
+
+}
